+++ src/Pdf.php
@@ -1,104 +1,58 @@
 <?php
 
-namespace Spatie\PdfToImage;
+namespace Drenso\PdfToImage;
 
-use Imagick;
-use Spatie\PdfToImage\Exceptions\InvalidFormat;
-use Spatie\PdfToImage\Exceptions\PageDoesNotExist;
-use Spatie\PdfToImage\Exceptions\PdfDoesNotExist;
+use Drenso\PdfToImage\Enum\ExportFormatEnum;
+use Drenso\PdfToImage\Exceptions\PageDoesNotExist;
+use Drenso\PdfToImage\Exceptions\PdfDoesNotExist;
+use GdImage;
+use Random\Engine\Secure;
 
 class Pdf
 {
-    protected $pdfFile;
+    protected string $cacheDir;
+    protected ?int $width = null;
+    protected ?ExportFormatEnum $outputFormat = null;
+    protected int $page = 1;
+    protected int $compressionQuality = -1;
+    private ?int $numberOfPages = null;
 
-    protected $resolution = 144;
-
-    protected $outputFormat = 'jpg';
-
-    protected $page = 1;
-
-    public $imagick;
-
-    protected $validOutputFormats = ['jpg', 'jpeg', 'png'];
-
-    protected $layerMethod = Imagick::LAYERMETHOD_FLATTEN;
-
-    protected $colorspace;
-
-    protected $compressionQuality;
-
-    protected $thumbnailWidth;
-
-    private $numberOfPages = null;
-
-    public function __construct(string $pdfFile)
+    public function __construct(private readonly string $pdfFile, protected readonly int $resolution = 144)
     {
         if (! file_exists($pdfFile)) {
             throw new PdfDoesNotExist("File `{$pdfFile}` does not exist");
         }
 
-        $this->pdfFile = $pdfFile;
-
-        $this->imagick = new Imagick();
-
-        $this->imagick->readImage($this->pdfFile);
+        // Convert to PNG files, so GD can be used for the following processing
+        $this->cacheDir = '/tmp/pdftoimage/' . bin2hex((new Secure())->generate());
+        @mkdir($this->cacheDir, recursive: true);
+        exec(sprintf('gs -dSAFER -dBATCH -sDEVICE=png16m -dTextAlphaBits=4 -dGraphicsAlphaBits=4 -r%s -sOutputFile=%s %s', $this->resolution, $this->cacheDir . '/%03d.png', $this->pdfFile));
     }
 
-    public function setResolution(int $resolution)
+    public function setWidth(int $width): self
     {
-        $this->resolution = $resolution;
+        $this->width = $width;
 
         return $this;
     }
 
-    public function setOutputFormat(string $outputFormat)
+    public function setOutputFormat(ExportFormatEnum $outputFormat): self
     {
-        if (! $this->isValidOutputFormat($outputFormat)) {
-            throw new InvalidFormat("Format {$outputFormat} is not supported");
-        }
-
         $this->outputFormat = $outputFormat;
 
         return $this;
     }
 
-    public function getOutputFormat(): string
+    public function getOutputFormat(): ExportFormatEnum
     {
         return $this->outputFormat;
     }
 
-    /**
-     * Sets the layer method for Imagick::mergeImageLayers()
-     * If int, should correspond to a predefined LAYERMETHOD constant.
-     * If null, Imagick::mergeImageLayers() will not be called.
-     *
-     * @param int|null
-     *
-     * @return $this
-     *
-     * @throws \Spatie\PdfToImage\Exceptions\InvalidLayerMethod
-     *
-     * @see https://secure.php.net/manual/en/imagick.constants.php
-     * @see Pdf::getImageData()
-     */
-    public function setLayerMethod(?int $layerMethod)
-    {
-        $this->layerMethod = $layerMethod;
-
-        return $this;
-    }
-
-    public function isValidOutputFormat(string $outputFormat): bool
-    {
-        return in_array($outputFormat, $this->validOutputFormats);
-    }
-
-    public function setPage(int $page)
+    public function setPage(int $page): self
     {
         if ($page > $this->getNumberOfPages() || $page < 1) {
             throw new PageDoesNotExist("Page {$page} does not exist");
         }
-
         $this->page = $page;
 
         return $this;
@@ -107,7 +61,8 @@ class Pdf
     public function getNumberOfPages(): int
     {
         if ($this->numberOfPages === null) {
-            $this->numberOfPages = $this->imagick->getNumberImages();
+            $files = scandir($this->cacheDir);
+            $this->numberOfPages = count(array_filter($files, fn (string $filename) => str_ends_with($filename, '.png')));
         }
 
         return $this->numberOfPages;
@@ -116,12 +71,11 @@ class Pdf
     public function saveImage(string $pathToImage): bool
     {
         if (is_dir($pathToImage)) {
-            $pathToImage = rtrim($pathToImage, '\/').DIRECTORY_SEPARATOR.$this->page.'.'.$this->outputFormat;
+            $pathToImage = rtrim($pathToImage, '\/').DIRECTORY_SEPARATOR.$this->page.$this->outputFormat->getExtension();
         }
-
         $imageData = $this->getImageData($pathToImage);
 
-        return file_put_contents($pathToImage, $imageData) !== false;
+        return $this->outputFormat->export($imageData, $pathToImage, $this->compressionQuality);
     }
 
     public function saveAllPagesAsImages(string $directory, string $prefix = ''): array
@@ -134,102 +88,41 @@ class Pdf
 
         return array_map(function ($pageNumber) use ($directory, $prefix) {
             $this->setPage($pageNumber);
-
             $destination = "{$directory}/{$prefix}{$pageNumber}.{$this->outputFormat}";
-
             $this->saveImage($destination);
 
             return $destination;
         }, range(1, $numberOfPages));
     }
 
-    public function getImageData(string $pathToImage): Imagick
+    public function getImageData(string $pathToImage): GdImage
     {
-        /*
-         * Reinitialize imagick because the target resolution must be set
-         * before reading the actual image.
-         */
-        $this->imagick = new Imagick();
+        $this->outputFormat ??= ExportFormatEnum::fromFileName($pathToImage);
+        $pageName = $this->cacheDir . sprintf('/%03d.png', $this->page);
+        $originalImage = imagecreatefrompng($pageName);
 
-        $this->imagick->setResolution($this->resolution, $this->resolution);
-
-        if ($this->colorspace !== null) {
-            $this->imagick->setColorspace($this->colorspace);
+        if ($this->width === null) {
+            return $originalImage;
         }
 
-        if ($this->compressionQuality !== null) {
-            $this->imagick->setCompressionQuality($this->compressionQuality);
+        $imageSize = getimagesize($pageName);
+        // Never grow the image
+        if ($imageSize[0] < $this->width) {
+            return $originalImage;
         }
+        // Calculate scaled height
+        $newHeight = $imageSize[1] * ($this->width / $imageSize[0]);
+        // Resize in new image
+        $resizedImage = imagecreatetruecolor($this->width, $newHeight);
+        imagecopyresampled($resizedImage, $originalImage, 0, 0, 0, 0, $this->width, $newHeight, $imageSize[0], $imageSize[1]);
 
-        if (filter_var($this->pdfFile, FILTER_VALIDATE_URL)) {
-            return $this->getRemoteImageData($pathToImage);
-        }
-
-        $this->imagick->readImage(sprintf('%s[%s]', $this->pdfFile, $this->page - 1));
-
-        if (is_int($this->layerMethod)) {
-            $this->imagick = $this->imagick->mergeImageLayers($this->layerMethod);
-        }
-
-        if ($this->thumbnailWidth !== null) {
-            $this->imagick->thumbnailImage($this->thumbnailWidth, 0);
-        }
-
-        $this->imagick->setFormat($this->determineOutputFormat($pathToImage));
-
-        return $this->imagick;
+        return $resizedImage;
     }
 
-    public function setColorspace(int $colorspace)
-    {
-        $this->colorspace = $colorspace;
-
-        return $this;
-    }
-
-    public function setCompressionQuality(int $compressionQuality)
+    public function setCompressionQuality(int $compressionQuality): self
     {
         $this->compressionQuality = $compressionQuality;
 
         return $this;
-    }
-
-    public function width(int $thumbnailWidth)
-    {
-        $this->thumbnailWidth = $thumbnailWidth;
-
-        return $this;
-    }
-
-    protected function getRemoteImageData(string $pathToImage): Imagick
-    {
-        $this->imagick->readImage($this->pdfFile);
-
-        $this->imagick->setIteratorIndex($this->page - 1);
-
-        if (is_int($this->layerMethod)) {
-            $this->imagick = $this->imagick->mergeImageLayers($this->layerMethod);
-        }
-
-        $this->imagick->setFormat($this->determineOutputFormat($pathToImage));
-
-        return $this->imagick;
-    }
-
-    protected function determineOutputFormat(string $pathToImage): string
-    {
-        $outputFormat = pathinfo($pathToImage, PATHINFO_EXTENSION);
-
-        if ($this->outputFormat != '') {
-            $outputFormat = $this->outputFormat;
-        }
-
-        $outputFormat = strtolower($outputFormat);
-
-        if (! $this->isValidOutputFormat($outputFormat)) {
-            $outputFormat = 'jpg';
-        }
-
-        return $outputFormat;
     }
 }